FHIR_COMBINED/FHIR_LLM_UA/backend/requirements.txt
accelerate==1.10.1
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1753980460404/work
anaconda-auth @ file:///croot/anaconda-cloud-auth-split_1747863777058/work
anaconda-cli-base @ file:///croot/anaconda-cli-base_1741369443553/work
annotated-types @ file:///work/perseverance-python-buildout/croot/annotated-types_1728386958747/work
anyio==4.11.0
archspec @ file:///croot/archspec_1709217642129/work
bitsandbytes==0.48.1
boltons @ file:///croot/boltons_1751383775419/work
brotlicffi @ file:///croot/brotlicffi_1736182461069/work
certifi @ file:///croot/certifi_1754570635119/work/certifi
cffi @ file:///croot/cffi_1736182485317/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click @ file:///croot/click_1754477934456/work
colorama @ file:///work/perseverance-python-buildout/croot/colorama_1728388434940/work
conda @ file:///croot/conda_1754469510968/work/conda-src
conda-anaconda-telemetry @ file:///croot/conda-anaconda-telemetry_1755883788794/work
conda-anaconda-tos @ file:///croot/conda-anaconda-tos_1755123313635/work
conda-content-trust @ file:///work/perseverance-python-buildout/croot/conda-content-trust_1728487058234/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1745607008911/work/src
conda-package-handling @ file:///croot/conda-package-handling_1731369017509/work
conda_package_streaming @ file:///croot/conda-package-streaming_1751307506088/work
cryptography @ file:///croot/cryptography_1753084572505/work
distro @ file:///work/perseverance-python-buildout/croot/distro_1728396110052/work
elastic-transport==8.17.1
elasticsearch==8.19.2
fastapi==0.119.0
filelock==3.20.0
frozendict @ file:///work/perseverance-python-buildout/croot/frozendict_1728497542215/work
fsspec==2025.9.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.1.10
huggingface-hub==0.35.3
idna @ file:///work/perseverance-python-buildout/croot/idna_1728385935861/work
jaraco.classes @ file:///croot/jaraco.classes_1755516327832/work
jaraco.context @ file:///io/perseverance-python-buildout/croot/jaraco.context_1731706750665/work
jaraco.functools @ file:///croot/jaraco.functools_1740408111753/work
jeepney @ file:///tmp/build/80754af9/jeepney_1627537048313/work
Jinja2==3.1.6
joblib==1.5.2
jsonpatch @ file:///work/perseverance-python-buildout/croot/jsonpatch_1728399595941/work
jsonpointer @ file:///croot/jsonpointer_1753788433671/work
keyring @ file:///croot/keyring_1745484691218/work
libmambapy @ file:///croot/mamba-split_1734469461757/work/libmambapy
markdown-it-py @ file:///work/perseverance-python-buildout/croot/markdown-it-py_1728387994477/work
MarkupSafe==3.0.3
mdurl @ file:///work/perseverance-python-buildout/croot/mdurl_1728387286143/work
menuinst @ file:///croot/menuinst_1753464582829/work
more-itertools @ file:///work/perseverance-python-buildout/croot/more-itertools_1728388668839/work
mpmath==1.3.0
mysql-connector-python==9.5.0
networkx==3.5
numpy==2.3.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
packaging @ file:///croot/packaging_1753775359799/work
pillow==12.0.0
pkce @ file:///work/perseverance-python-buildout/croot/pkce_1728388093137/work
platformdirs @ file:///croot/platformdirs_1744273042065/work
pluggy @ file:///croot/pluggy_1733169602837/work
psutil==7.1.0
pycosat @ file:///croot/pycosat_1736868416091/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic @ file:///croot/pydantic_1750768261732/work
pydantic-settings @ file:///io/buildout/croot/pydantic-settings_1735843225507/work
pydantic_core @ file:///croot/pydantic-core_1750754736071/work
Pygments @ file:///croot/pygments_1744664109463/work
PyJWT @ file:///io/buildout/croot/pyjwt_1739558969999/work
PyMySQL==1.1.2
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1728386193338/work
python-dateutil==2.9.0.post0
python-dotenv @ file:///croot/python-dotenv_1745610241282/work
PyYAML==6.0.3
readchar @ file:///work/perseverance-python-buildout/croot/readchar_1728397613245/work
regex==2025.9.18
requests @ file:///croot/requests_1750426329888/work
rich @ file:///croot/rich_1732638981168/work
ruamel.yaml @ file:///croot/ruamel.yaml_1745960305322/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1745937152469/work
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
SecretStorage @ file:///work/perseverance-python-buildout/croot/secretstorage_1728387678881/work
semver @ file:///croot/semver_1736540782504/work
sentence-transformers==5.1.2
setuptools==78.1.1
shellingham @ file:///work/perseverance-python-buildout/croot/shellingham_1728395303315/work
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
tomli @ file:///croot/tomli_1753774587605/work
torch==2.9.0
tqdm @ file:///croot/tqdm_1738943501192/work
transformers==4.57.1
triton==3.5.0
truststore @ file:///croot/truststore_1753907171582/work
typer @ file:///work/perseverance-python-buildout/croot/typer_1728397851651/work
typing-inspection @ file:///croot/typing-inspection_1746023470701/work
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
urllib3 @ file:///croot/urllib3_1750775463400/work
uvicorn==0.38.0
wheel==0.45.1
zstandard @ file:///croot/zstandard_1731356346222/work
